Broadstreet ad zones updated
I created 3 ad zones specific to bills site for a billboard, any kind of interstitial, and a keyword-based zone. BS is designed so that ads find where they're supposed to go, regardless of where the ad zone scripts are placed (although I am under impression that the watch code has to come first).